endianness: Cast bswap operand to fixed width in MSVC macros
The MSVC RUINT{16,32,64}_BSWAP macros expanded to bare
_byteswap_ushort / _ulong / _uint64 calls, so passing a wider
value (e.g. RUINT32_BSWAP(v64) in /rendianness/bswap, exercising
the low-32 truncation) triggered C4244 at the call site. The
GCC/builtin sibling variants in the same file already include
the matching internal cast -- (rint32) before __builtin_bswap32
etc. Add the analogous (unsigned short / unsigned long /
unsigned __int64) cast to the MSVC variants so both backends
silently truncate the operand consistently.

+++ include/rlib/types/rendianness.h
@@ -39,9 +39,9 @@ R_BEGIN_DECLS
 #define RUINT32_BSWAP(val) ((ruint32) __builtin_bswap32 ((rint32) (val)))
 #define RUINT64_BSWAP(val) ((ruint64) __builtin_bswap64 ((rint64) (val)))
 #elif defined(_MSC_VER)
-#define RUINT16_BSWAP(val) (_byteswap_ushort (val))
-#define RUINT32_BSWAP(val) (_byteswap_ulong (val))
-#define RUINT64_BSWAP(val) (_byteswap_uint64 (val))
+#define RUINT16_BSWAP(val) (_byteswap_ushort ((unsigned short)(val)))
+#define RUINT32_BSWAP(val) (_byteswap_ulong ((unsigned long)(val)))
+#define RUINT64_BSWAP(val) (_byteswap_uint64 ((unsigned __int64)(val)))
 #else
 #define RUINT16_BSWAP(val) ((ruint16) (((ruint16)(val) >> 8) | ((ruint16)(val) << 8)))
 #define RUINT32_BSWAP(val) ((ruint32) (                                   \
